move SectionTable into RightPane folder
