+++ src/components/RightPane/SectionTable/OpenSpotAlertPopover.js
@@ -3,9 +3,9 @@ import PropTypes from 'prop-types';
 import { withStyles } from '@material-ui/core/styles';
 import InputMask from 'react-input-mask';
 import { Button, Popover, TextField, Typography } from '@material-ui/core';
-import { openSnackbar } from '../../actions/AppStoreActions';
-import { REGISTER_NOTIFICATIONS_ENDPOINT } from '../../api/endpoints';
-import dispatcher from '../../dispatcher';
+import { openSnackbar } from '../../../actions/AppStoreActions';
+import { REGISTER_NOTIFICATIONS_ENDPOINT } from '../../../api/endpoints';
+import dispatcher from '../../../dispatcher';
 
 const phoneNumberRegex = RegExp(/\d{10}/);
