+++ src/components/RightPane/SectionTable/static/locations.json
