+++ src/components/RightPane/SectionTable/static/restrictionsMapping.json
